Extract SC monitor class
